requirements.txt
accelerate==0.25.0
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
beartype==0.16.4
cmake==3.27.6
colored==2.2.4
contourpy==1.1.1
cycler==0.12.1
datasets==2.16.1
dill==0.3.7
distro==1.8.0
evaluate==0.4.1
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.6.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.2
humanfriendly==10.0
humanize==4.9.0
kiwisolver==1.4.5
matplotlib==3.8.0
multidict==6.0.4
multiprocess==0.70.15
ninja==1.11.1.1
pandas==2.1.1
protobuf==4.25.1
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==13.0.0
pyarrow-hotfix==0.6
pynvml==11.5.0
pyparsing==3.1.1
regex==2023.10.3
responses==0.18.0
safetensors==0.4.1
seaborn==0.13.0
sentencepiece==0.1.99
slurm-gpustat==0.0.15
tabulate==0.9.0
termcolor==2.4.0
tokenizers==0.15.0
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.37.2
triton==2.1.0
tzdata==2023.3
xxhash==3.4.1
yarl==1.9.2
wandb==0.19.3
cupy-cuda12x==13.3.0